Fixed a bug w.r.t. looking up data from the LRU and static entries for delta2
Added a 'done' call so that stats, etc. can be finalized for compressors
which wish such a thing to happen. This was used to print out the final
letter-frequency counts at the end of the run so that the huffman codes
could be optimized.

Regenerated the letter-frequency-counts based on the data checked into
the http_samples repo

Added canonical huffman coding and some (dead) code (which I'll need to
rip out soon) which would aid in decoding speed were the non-naive
decode used.

Added printouts of the instructions and hex of the serialized data,
suitable for using as examples in a specification...